youtube-backend/requirements.txt
argcomplete==2.0.0
asn1crypto==1.5.1
awscli==1.25.76
beautifulsoup4==4.11.1
boto3==1.24.56
botocore==1.27.75
certifi==2022.6.15
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
docutils==0.16
durationpy==0.5
Flask==2.2.2
hjson==3.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.5
packaging==21.3
pandas==1.4.2
pem==21.2.0
placebo==0.9.0
pyasn1==0.4.8
pycryptodome==3.15.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2021.3
PyYAML==5.4.1
redshift-connector==2.0.902
requests==2.26.0
rsa==4.7.2
s3transfer==0.6.0
scramp==1.4.4
six==1.16.0
soupsieve==2.3.2.post1
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.0
troposphere==4.1.0
urllib3==1.26.11
Werkzeug==2.2.2
wsgi-request-logger==0.4.6
zappa==0.55.0
zipp==3.8.1
Flask-Cors==4.0.0
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
openai==0.28.1
load_dotenv==0.1.0
